[Release] | 3.0.1 (#42)
## [3.0.1] - Beetle

### Added
- `shouldShowResultScreen` to `@objc TrueLayerSinglePaymentPreferences`
and `@objc TrueLayerMandatePreferences` objects (default value is
`true`). When this is enabled, a result screen is displayed at the end
of the payment flow. When the user is redirected back from the bank, it
is recommended to re-invoke the TrueLayer SDK to display the result
screen, to show the user the status of their payment or mandate.
- `maximumResultScreenTimeout` to `@objc
TrueLayerSinglePaymentPreferences` and `@objc
TrueLayerMandatePreferences` objects. This is the maximum timeout for
the payment or mandate result screen, until a final status. Once
reached, the user is shown a button to dismiss the SDK and return to
your app.