[Release] | 3.1.0 (#44)
## [3.1.0] - ResultShown

### Added 
- `resultShown` to `processSinglePayment` and `processMandate`
completion handlers. This `enum` informs which result was shown on the
result screen, if it was shown (`shouldShowResultScreen` is set to
`true`.) If not, `ResultShown.noneShown` will be returned.
- Search for countries by their 2 and 3-letter country codes.
- Error shown for additional inputs when no options are available.

### Changed
- New `processSinglePayment` and `processMandate` methods. The object
returned in the completion handler is now of type
`TrueLayer.Payments.Models.SinglePayment.ProcessResult` and
`TrueLayer.Payments.Models.Mandate.ProcessResult`. These are `enum`s
with associated objects that contain `state` and `error` properties, as
you would access previously.
- Improved branch searching to better match intended branch names.
- Default colours for the primary and secondary buttons.

---------

Co-authored-by: Daven Gomes <daven.gomes@truelayer.com>